Added evil hangman project
Completed in Spring 2018, Implementation of evil hangman game in C. UML's Computing 2 semester long project, it can be done in significantly less code in other languages like C++. The point of doing it in C was to teach students proper management of dynamically allocated memory as well as learning the data structures involved. Each week we had tasks to complete, and checking of memory leaks was done using valgrind.